LayoutTests/imported/w3c/web-platform-tests/css/css-borders/corner-shape/corner-top-left-valid-expected.txt


FAIL e.style['corner-top-left'] = "round 30%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['corner-top-left'] = "10px bevel" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['corner-top-left'] = "normal" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['corner-top-left'] = "4px 2% round" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['corner-top-left'] = "superellipse(-0.5) 30% 10px" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['corner-top-left'] = "10px bevel" should set border-top-left-radius assert_equals: border-top-left-radius should be canonical expected "10px" but got ""
FAIL e.style['corner-top-left'] = "10px bevel" should set corner-top-left-shape assert_equals: corner-top-left-shape should be canonical expected "bevel" but got ""
FAIL e.style['corner-top-left'] = "10px bevel" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-top-left'] = "normal" should set border-top-left-radius assert_equals: border-top-left-radius should be canonical expected "0px" but got ""
FAIL e.style['corner-top-left'] = "normal" should set corner-top-left-shape assert_equals: corner-top-left-shape should be canonical expected "round" but got ""
FAIL e.style['corner-top-left'] = "normal" should not set unrelated longhands assert_true: expected true got false
